+++ src/main/java/com/productdock/book/BookService.java
@@ -16,7 +16,7 @@ public record BookService(BookRepository bookRepository,
                 .toList();
     }
 
-    public Integer countAllBooks() {
-        return Math.toIntExact(bookRepository.count());
+    public long countAllBooks() {
+        return bookRepository.count();
     }
 }
